requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
ultralytics==8.0.200
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.1.0
langchain==0.1.0
langchain-openai==0.0.2.post1
pydantic==2.5.3